Removed versions.cfg and Makefile.
The versions.cfg was only used in the Makefile to get a version for 'black'.
This failed because the command was wrong, trying to install a package called `install`:

```
bin/pip install pip install black==$(awk '/^black =/{print $NF}' versions.cfg)
Requirement already satisfied: pip in ./lib/python3.13/site-packages (25.1.1)
ERROR: Could not find a version that satisfies the requirement install (from versions: none)
```

Also, the Makefile was using Python 3.9, which is not supported on Plone 6.